refactor : controller 리팩토링 feat : DB 및 DTO 추가 및 수정

+++ src/main/java/com/example/reward_monitoring/mission/saveMsn/controller/SaveMsnController.java
@@ -1,8 +1,9 @@
 package com.example.reward_monitoring.mission.saveMsn.controller;
 
-import com.example.reward_monitoring.mission.saveMsn.dto.SaveMsnEditDto;
-import com.example.reward_monitoring.mission.saveMsn.dto.SaveMsnReadDto;
-import com.example.reward_monitoring.mission.saveMsn.dto.SaveMsnSearchDto;
+import com.example.reward_monitoring.general.member.entity.Member;
+import com.example.reward_monitoring.general.member.model.Auth;
+import com.example.reward_monitoring.general.member.repository.MemberRepository;
+import com.example.reward_monitoring.mission.saveMsn.dto.*;
 import com.example.reward_monitoring.mission.saveMsn.entity.SaveMsn;
 import com.example.reward_monitoring.mission.saveMsn.repository.SaveMsnRepository;
 import com.example.reward_monitoring.mission.saveMsn.service.SaveMsnService;
@@ -12,6 +13,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
@@ -29,21 +31,57 @@ import java.util.List;
 
 @Controller
 @Tag(name = "saveMsn", description = "저장미션 API")
+@RequestMapping(value = "/Mission/sightseeingList")
 public class SaveMsnController {
     @Autowired
     private SaveMsnRepository saveMsnRepository;
-
     @Autowired
     private SaveMsnService saveMsnService;
+    @Autowired
+    private MemberRepository memberRepository;
+
 
     @Operation(summary = "저장미션 정보 수정", description = "저장미션 정보를 수정합니다")
     @Parameter(name = "idx", description = "수정할 저장미션의 IDX")
-    @PostMapping("/saveMsn/edit/{idx}")
+    @PostMapping("/edit/{idx}")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "성공적으로 수정됨"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
             @ApiResponse(responseCode = "500", description = "일치하는 미션을 찾을 수 없음")
     })
-    public ResponseEntity<SaveMsn> edit(@PathVariable int idx, @RequestBody SaveMsnEditDto dto, HttpServletResponse response){
+    public ResponseEntity<SaveMsn> edit(HttpSession session,
+                                        @PathVariable int idx,
+                                        @RequestPart(value ="file",required = false)MultipartFile multipartFile,
+                                        @RequestPart(value="dto",required = true) SaveMsnEditDto dto,
+                                        HttpServletResponse response){
+
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+        Member member =memberRepository.findById(sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+        if(member.getAuthSaveMsn()== Auth.READ) // 읽기 권한만 존재할경우
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+        if(multipartFile != null && !multipartFile.isEmpty()){
+            try {
+                byte[] fileContent = multipartFile.getBytes();
+                String originalFilename = multipartFile.getOriginalFilename();
+                dto.setImageData(fileContent);
+                dto.setImageName(originalFilename);
+            } catch (IOException e) {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+            }
+        }
+
 
         SaveMsn edited = saveMsnService.edit(idx,dto);
 
@@ -56,26 +94,74 @@ public class SaveMsnController {
     }
     
     @Operation(summary = "저장미션 생성", description = "저장미션 정보를 생성합니다")
-    @PostMapping("/saveMsn/add")
+    @PostMapping("/add")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "성공적으로 수정됨"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
             @ApiResponse(responseCode = "500", description = "서버 오류로 미션생성이 되지않음")
     })
-    public ResponseEntity<Void> add(@RequestBody SaveMsnReadDto dto){
+    public ResponseEntity<Void> add(HttpSession session,
+                                    @RequestPart(value ="file",required = false)MultipartFile multipartFile,
+                                    @RequestPart(value ="dto",required = true) SaveMsnReadDto dto){
+
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+        Member member =memberRepository.findById( sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+        if(member.getAuthSaveMsn()== Auth.READ) // 읽기 권한만 존재할경우
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+
+        if(multipartFile != null && !multipartFile.isEmpty()){
+            try {
+                byte[] fileContent = multipartFile.getBytes();
+                String originalFilename = multipartFile.getOriginalFilename();
+                dto.setImageData(fileContent);
+                dto.setImageName(originalFilename);
+            } catch (IOException e) {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+            }
+        }
+
         SaveMsn created = saveMsnService.add(dto);
         if(created == null)
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         saveMsnRepository.save(created);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
+
+
     @Operation(summary = "저장미션 정보 요청", description = "IDX와 일치하는 미션정보를 반환합니다")
     @Parameter(name = "idx", description = "관리자 IDX")
-    @GetMapping("saveMsn/{idx}") 
+    @GetMapping("{idx}")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "미션 검색 완료 "),
-            @ApiResponse(responseCode = "204", description = "일치하는 미션을 찾을 수 없음")
+            @ApiResponse(responseCode = "204", description = "일치하는 미션을 찾을 수 없음"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
     })
-    public ResponseEntity<SaveMsn> getSaveMsn(@PathVariable int idx){
+    public ResponseEntity<SaveMsn> getSaveMsn(HttpSession session,@PathVariable int idx){
+
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+        Member member =memberRepository.findById(sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
         SaveMsn target = saveMsnService.getSaveMsn(idx);
         return (target != null) ?
                 ResponseEntity.status(HttpStatus.OK).body(target):
@@ -83,22 +169,53 @@ public class SaveMsnController {
     }
 
     @Operation(summary = "전체 미션정보 요청", description = "전체 미션 정보를 반환합니다")
-    @GetMapping("/saveMsn/saveMsns")  //전체 광고주 리스트 반환
+    @GetMapping("/saveMsns")  //전체 광고주 리스트 반환
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "검색 완료"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음")
     })
-    public ResponseEntity<List<SaveMsn>> getSaveMsns(){
+    public ResponseEntity<List<SaveMsn>> getSaveMsns(HttpSession session){
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+        Member member =memberRepository.findById(sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+
         return ResponseEntity.status(HttpStatus.OK).body(saveMsnService.getSaveMsns());
     }
 
     @Operation(summary = "미션 삭제", description = "IDX와 일치하는 단일 미션정보를 삭제합니다")
     @Parameter(name = "idx", description = "미션 IDX")
-    @DeleteMapping("saveMsn/delete/{idx}")
+    @DeleteMapping("/delete/{idx}")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "204", description = "삭제 완료 "),
-            @ApiResponse(responseCode = "400", description = "일치하는 미션이 없음")
+            @ApiResponse(responseCode = "400", description = "일치하는 미션이 없음"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음")
     })
-    public ResponseEntity<String> delete(@PathVariable int idx)throws IOException {
+    public ResponseEntity<String> delete(HttpSession session,@PathVariable int idx)throws IOException {
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+        Member member =memberRepository.findById( sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+        if(member.getAuthSaveMsn()== Auth.READ) // 읽기 권한만 존재할경우
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
 
         SaveMsn deleted  = saveMsnService.delete(idx);
         return (deleted != null) ?
@@ -106,22 +223,92 @@ public class SaveMsnController {
                 ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
 
     }
+
     @Operation(summary = "저장미션 검색", description = "조건에 맞는 저장미션을 검색합니다")
-    @PostMapping("/saveMsn/search")
+    @PostMapping("/search")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "검색 완료(조건에 맞는결과가없을경우 빈 리스트 반환)"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
             @ApiResponse(responseCode = "500", description = "검색 중 예기치않은 오류발생")
     })
-    public ResponseEntity<List<SaveMsn>> searchMember(@RequestBody SaveMsnSearchDto dto){
+    public ResponseEntity<List<SaveMsn>> searchMember(HttpSession session,@RequestBody SaveMsnSearchDto dto){
+
+        Member sessionMember= (Member)session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+        Member member =memberRepository.findById(sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
         List<SaveMsn> result = saveMsnService.searchSaveMsn(dto);
         return (result != null) ?
                 ResponseEntity.status(HttpStatus.OK).body(result): // 일치하는 결과가 없을경우 빈 리스트 반환
                 ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
     }
+    @Operation(summary = "저장미션 현재 리스트 소진량 검색", description = "현재 리스트 소진량 페이지에서 조건에 맞는 저장미션을 검색합니다")
+    @PostMapping("/saveMsnByConsumed")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "검색 완료(조건에 맞는결과가없을경우 빈 리스트 반환)"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
+            @ApiResponse(responseCode = "500", description = "검색 중 예기치않은 오류발생")
+    })
+    public ResponseEntity<List<SaveMsn>> searchSaveMsnByConsumed(HttpSession session, @RequestBody SaveMsnSearchDto dto){
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
 
-    protected int ind() {
-        int in = 50;
-        return in;
+        Member member =memberRepository.findById( sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+
+        if(member.isNauthCurSave()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+
+        List<SaveMsn> result = saveMsnService.searchSaveMsnByConsumed(dto);
+        return (result != null) ?
+                ResponseEntity.status(HttpStatus.OK).body(result): // 일치하는 결과가 없을경우 빈 리스트 반환
+                ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+    }
+    
+    @Operation(summary = "모든 미션 종료", description = "!!DB의 모든 미션을 종료하고 비노출처리합니다!!")
+    @GetMapping("/endAll")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "모든 미션 종료 처리"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
+            @ApiResponse(responseCode = "500", description = "작업 중 예기치않은 오류발생")
+    })
+    public ResponseEntity<Void> allMissionEnd(HttpSession session){
+        Member sessionMember= (Member) session.getAttribute("member");
+        if(sessionMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } // 세션만료
+
+        Member member =memberRepository.findById( sessionMember.getId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }//데이터 없음
+
+        if(member.isNauthSaveMsn()) // 비권한 활성화시
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        if(member.getAuthSaveMsn()== Auth.READ) // 읽기 권한만 존재할경우
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
+
+        boolean result = saveMsnService.allMissionEnd();
+
+        return (result) ?
+                ResponseEntity.status(HttpStatus.OK).build():
+                ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
     }
 
     @Operation(summary = "엑셀 다운로드", description = "저장미션 리스트 엑셀파일을 다운로드합니다")
@@ -166,4 +353,72 @@ public class SaveMsnController {
                 ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
     }
 
+    @Operation(summary = "미션 사용여부 변경", description = "미션 사용여부를 변경합니다")
+    @PostMapping("/reEngagement/{idx}")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
+            @ApiResponse(responseCode = "500", description = "")
+    })
+    public ResponseEntity<Void> changeMissionActive(HttpSession session, @PathVariable int idx , @RequestBody SaveMsnActiveDto dto)throws IOException {
+
+
+        boolean result = saveMsnService.changeMissionActive(idx,dto);
+
+        return (result) ?
+                ResponseEntity.status(HttpStatus.OK).build():
+                ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+    }
+
+    @Operation(summary = "미션 노출여부 변경", description = "미션 노출여부를 변경합니다.")
+    @PostMapping("/expose/{idx}")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
+            @ApiResponse(responseCode = "500", description = "엑셀파일의 문제로 인한 데이터 삽입 실패")
+    })
+    public ResponseEntity<Void> changeMissionExpose(HttpSession session, @PathVariable int idx , @RequestBody SaveMsnExposeDto dto)throws IOException {
+
+        boolean result = saveMsnService.changeMissionExpose(idx,dto);
+
+        return (result) ?
+                ResponseEntity.status(HttpStatus.OK).build():
+                ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+    }
+
+    @Operation(summary = "가능일 변경", description = "가능일을 변경합니다")
+    @PostMapping("/{idx}/changeDay")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "401", description = "세션이 없거나 만료됨"),
+            @ApiResponse(responseCode = "403", description = "권한없음"),
+            @ApiResponse(responseCode = "500", description = "엑셀파일의 문제로 인한 데이터 삽입 실패")
+    })
+    public ResponseEntity<Void> changeAbleDay(HttpSession session, SaveMsnAbleDayDto dto, @PathVariable int idx)throws IOException {
+        boolean result = saveMsnService.changeAbleDay(dto,idx);
+
+        return (result) ?
+                ResponseEntity.status(HttpStatus.OK).build():
+                ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+    }
+
+
+    @GetMapping("")
+    public String quizList(){
+        return "searchList";
+    }
+
+    @GetMapping("/Mission/quizCurrentList")
+    public void quizCurrentList(HttpSession session, HttpServletResponse response) throws IOException {
+//        // 권한 검사 로직 (예시)
+//        if (!hasPermission()) {
+//            response.sendError(HttpStatus.UNAUTHORIZED.value(), "Unauthorized"); // 401 오류를 반환
+//            return;
+//        }
+//
+//        // 권한이 있을 경우 다른 페이지로 리다이렉트
+//        response.sendRedirect("/anotherPage"); // 이동할 URL
+    }
 }
